backend/requirements.txt
flask
flask-cors
langchain
transformers
torch
python-dotenv